Correcting errors& fixing code after testing [3]

+++ src/main/java/com/libertywallet/dto/BudgetDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 
 @Data
 public class BudgetDto {
-    private BigDecimal current_balance;
+    private int current_balance;
     private int amountLimit;
     private LocalDate start_date;
     private LocalDate end_date;
