+++ src/main/java/com/libertywallet/request/BudgetRequest.java
@@ -2,13 +2,11 @@ package com.libertywallet.request;
 
 
 import lombok.Data;
-
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class BudgetRequest {
-    BigDecimal current_balance;
+    int current_balance;
     int amountLimit;
     LocalDate start_date;
     LocalDate end_date;
