+++ src/main/java/com/libertywallet/request/TransactionRequest.java
@@ -3,14 +3,12 @@ package com.libertywallet.request;
 
 
 import lombok.Data;
-
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class TransactionRequest {
     Long categoryId;
-    BigDecimal amount;
+    int amount;
     String description;
     LocalDate date;
 }
